requirements.txt
flask==2.0.1
numpy==1.21.2
opencv-python==4.5.3
pillow==8.3.2
qrcode==7.3.1
pyzbar==0.1.8
werkzeug==2.0.2
ultralytics==8.0.0
pandas
fast-alpr>=0.1.0 